CommandsConfiguration and MysqlConfiguration
